+++ src/web/src/views/workspace/WSEditorCommandArgumentsContent/ArgNavBar.tsx
