+++ src/web/src/views/workspace/WSEditorCommandArgumentsContent/ArgumentNavigation.tsx
