+++ src/web/src/views/workspace/WSEditorCommandArgumentsContent/FlattenDialog.tsx
