+++ src/web/src/views/workspace/WSEditorCommandArgumentsContent/WSEditorCommandArgumentsContent.tsx
